backend/requirements.txt
# Backend Dependencies for Crypto Trade Logger
# =============================================
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database - Async SQLAlchemy with PostgreSQL
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
greenlet>=3.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# CORS and HTTP
python-multipart>=0.0.12

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
